backend/requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
bs4
python-dotenv
psycopg2
psycopg2-binary
apscheduler
flask_cors